Only use sha256 if fips is enabled
