+++ internal/internal_worker.go
@@ -28,8 +28,6 @@ package internal
 
 import (
 	"context"
-	"crypto/sha256"
-	"encoding/hex"
 	"errors"
 	"fmt"
 	"io"
@@ -1214,36 +1212,6 @@ func initBinaryChecksum() error {
 	defer binaryChecksumLock.Unlock()
 
 	return initBinaryChecksumLocked()
-}
-
-// callers MUST hold binaryChecksumLock before calling
-func initBinaryChecksumLocked() error {
-	if len(binaryChecksum) > 0 {
-		return nil
-	}
-
-	exec, err := os.Executable()
-	if err != nil {
-		return err
-	}
-
-	f, err := os.Open(exec)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		_ = f.Close() // error is unimportant as it is read-only
-	}()
-
-	h := sha256.New()
-	if _, err := io.Copy(h, f); err != nil {
-		return err
-	}
-
-	checksum := h.Sum(nil)
-	binaryChecksum = hex.EncodeToString(checksum[:])
-
-	return nil
 }
 
 func getBinaryChecksum() string {
